Set tags for cuda buildchain repositories (#123)
Add tags of the form "nb-N" for the rest of the cuda build chain
repositories in red-hat-data-services.  We did this for the
base images, add tags for minimal, pytorch, and tensorflow as
well so that the notebook source does not change for a release
because of a commit to the repository.