Add oauth-proxy authentication to BB exporter pod
This blackbox exporter can potentially be used to execute http queries
against any arbitrary endpoint. This change adds the oauth proxy in
front of the blackbox exporter service so that only authorized users can
interact with this service. It also configures the prometheus scrape
jobs which use the blackbox exporter to authenticate to the blackbox
exporter using an OpenShift service account's bearer token.