Delete tags for prebuilt images on upgrade
This change removes imagestream tags for prebuilt images
in the odh-deployer on a version upgrade. The operator
will create tags when it starts. This is to prevent the
accumulation of tags on multiple upgrades (the operator
will add a new tag, but not remove existing).

It also moves the buildchain.sh script to the end of the
init routine, so that the time to tag creation is as short
as possible.